Bot/requirements.txt
setuptools
discord.py
asyncio
aiosqlite
riotwatcher
python-dotenv
pantheon
matplotlib==3.7
opencv-python